+++ src/engine/compute.ts
@@ -1,7 +1,7 @@
 import type { BQLMutationBlock, EnrichedDataField } from '../types';
 import { getParamNames } from './helpers';
 
-export const compute = ({
+export const computeField = ({
 	currentThing,
 	fieldSchema,
 	mandatoryDependencies = false,
@@ -10,18 +10,28 @@ export const compute = ({
 	fieldSchema?: EnrichedDataField;
 	mandatoryDependencies?: boolean;
 }) => {
-	if (!fieldSchema || !fieldSchema.default || !fieldSchema.default.value) {
+	if (!fieldSchema || !fieldSchema.default || !('fn' in fieldSchema.default || 'value' in fieldSchema.default)) {
 		throw new Error('Virtual field: No field schema found, or wrongly configured');
 	}
-	const fn = fieldSchema.default.value;
-
-	const args = getParamNames(fn);
 
 	//check if all the args are in the entity, if not, throw a missing error with all the not present ones
-	const missingArgs = args.filter((arg) => !(arg in currentThing));
-	if (mandatoryDependencies && missingArgs.length) {
-		throw new Error(`Virtual field: Missing arguments ${missingArgs.join(', ')}`);
+
+	if (fieldSchema.default.type === 'value') {
+		return fieldSchema.default.value;
 	}
-	const computedValue = 'default' in fieldSchema ? fieldSchema.default?.value(currentThing) : undefined;
+
+	if (mandatoryDependencies) {
+		const { fn } = fieldSchema.default;
+
+		const args = getParamNames(fn);
+		const missingArgs = args.filter((arg) => !(arg in currentThing));
+
+		if (missingArgs.length) {
+			throw new Error(`Virtual field: Missing arguments ${missingArgs.join(', ')}`);
+		}
+	}
+	const computedValue = 'default' in fieldSchema ? fieldSchema.default?.fn(currentThing) : undefined;
 	return computedValue;
 };
+
+export const computeNode = () => {};