+++ src/types/requests/mutations.ts
@@ -1,3 +1,5 @@
+import type { EnrichedBormEntity, EnrichedBormRelation } from '../schema/enriched';
+import type { Schema } from '../symbols';
 import type { Filter } from './filters';
 
 type RequiredKey<T, K extends keyof T> = T & { [P in K]-?: T[P] };
@@ -11,7 +13,10 @@ export type BQLMutationBlock = {
 	$tempId?: string;
 	$op?: string;
 } & ({ $entity: string } | { $relation: string } | { $thing: string; $thingType: 'entity' | 'relation' }); // | { $attribute: string });
-export type FilledBQLMutationBlock = WithRequired<BQLMutationBlock, '$tempId' | '$op'>;
+export type FilledBQLMutationBlock = WithRequired<BQLMutationBlock, '$tempId' | '$op'> & {
+	$entity?: string;
+	[Schema]: EnrichedBormEntity | EnrichedBormRelation;
+};
 
 export type RawBQLMutation<T extends Record<string, any> = Record<string, any>> = (
 	| {