+++ src/types/symbols/index.ts
@@ -1,0 +1,1 @@
+export const Schema = Symbol.for('schema');